Move checking validity of argument to TaskLogic
