+++ src/main/java/seedu/sherpass/command/AddRecurringCommand.java
@@ -3,15 +3,11 @@ package seedu.sherpass.command;
 import seedu.sherpass.enums.Frequency;
 import seedu.sherpass.task.Task;
 import seedu.sherpass.task.TaskList;
-import seedu.sherpass.util.CommonLogic;
+import seedu.sherpass.task.TaskLogic;
 import seedu.sherpass.util.Storage;
 import seedu.sherpass.util.Ui;
 
 import java.time.LocalDateTime;
-
-import static seedu.sherpass.constant.Message.ERROR_EMPTY_DESCRIPTION_MESSAGE;
-import static seedu.sherpass.constant.Message.ERROR_INVALID_FREQUENCY_MESSAGE;
-import static seedu.sherpass.constant.Message.ERROR_START_AFTER_END_TIME_MESSAGE;
 
 public class AddRecurringCommand extends Command {
     private String taskDescription;
@@ -39,20 +35,9 @@ public class AddRecurringCommand extends Command {
         this.frequency = frequency;
     }
 
-    private String isValidArgument() {
-        if (taskDescription.isBlank()) {
-            return ERROR_EMPTY_DESCRIPTION_MESSAGE;
-        } else if (doOnStartDateTime.isAfter(doOnEndDateTime)) {
-            return ERROR_START_AFTER_END_TIME_MESSAGE;
-        } else if (frequency == null) {
-            return ERROR_INVALID_FREQUENCY_MESSAGE;
-        }
-        return "";
-    }
-
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
-        String error = isValidArgument();
+        String error = TaskLogic.checkValidAddArgument(taskDescription, doOnStartDateTime, doOnEndDateTime, frequency);
         if (!error.isBlank()) {
             ui.showToUser(error);
             return;
@@ -60,7 +45,7 @@ public class AddRecurringCommand extends Command {
         int identifier = taskList.generateIdentifier();
         Task newTask = new Task(identifier, taskDescription, null,
                 doOnStartDateTime, doOnEndDateTime, frequency);
-        LocalDateTime endDate = CommonLogic.getEndDateForRecurrence(doOnStartDateTime, frequency);
+        LocalDateTime endDate = TaskLogic.getEndDateForRecurrence(doOnStartDateTime, frequency);
 
         int count = 0;
         StringBuilder addedTaskString = new StringBuilder();
@@ -69,8 +54,8 @@ public class AddRecurringCommand extends Command {
             taskList.addTask(newTask);
             addedTaskString.append(newTask);
             addedTaskString.append("\n ");
-            doOnStartDateTime = CommonLogic.incrementDate(doOnStartDateTime, frequency);
-            doOnEndDateTime = CommonLogic.incrementDate(doOnEndDateTime, frequency);
+            doOnStartDateTime = TaskLogic.incrementDate(doOnStartDateTime, frequency);
+            doOnEndDateTime = TaskLogic.incrementDate(doOnEndDateTime, frequency);
             newTask = new Task(identifier, taskDescription, null, doOnStartDateTime, doOnEndDateTime, frequency);
         } while (newTask.getDoOnStartDateTime().isBefore(endDate));
