+++ src/main/java/seedu/sherpass/command/DeleteRecurringCommand.java
@@ -1,12 +1,9 @@
 package seedu.sherpass.command;
 
 import seedu.sherpass.task.TaskList;
+import seedu.sherpass.task.TaskLogic;
 import seedu.sherpass.util.Storage;
 import seedu.sherpass.util.Ui;
-
-import static seedu.sherpass.constant.Message.EMPTY_STRING;
-import static seedu.sherpass.constant.Message.ERROR_EMPTY_TASKLIST_MESSAGE;
-import static seedu.sherpass.constant.Message.ERROR_INVALID_INDEX_MESSAGE;
 
 public class DeleteRecurringCommand extends Command {
     private int index;
@@ -23,18 +20,11 @@ public class DeleteRecurringCommand extends Command {
         this.index = index;
     }
 
-    private String isValidArgument(TaskList taskList) {
-        if (taskList.getTasks().size() == 0) {
-            return ERROR_EMPTY_TASKLIST_MESSAGE;
-        } else if (!taskList.isTaskExist(index)) {
-            return ERROR_INVALID_INDEX_MESSAGE;
-        }
-        return EMPTY_STRING;
-    }
+
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
-        String error = isValidArgument(taskList);
+        String error = TaskLogic.checkValidDeleteArgument(taskList, index);
         if (!error.isBlank()) {
             ui.showToUser(error);
             return;
